Make UI

+++ app/src/main/java/ir/moonify/exchangeapp/model/CurrencyHolder.kt
@@ -1,3 +1,3 @@
 package ir.moonify.exchangeapp.model
 
-data class CurrencyHolder(val name: String, val value: Int)
+data class CurrencyHolder(val name: String, val value: Double)
